add-admin-review

+++ src/main/java/com/bravos/steak/store/entity/Game.java
@@ -6,7 +6,6 @@ import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @FieldDefaults(level = AccessLevel.PRIVATE)
@@ -29,7 +28,7 @@ public class Game {
     String name;
 
     @Column(precision = 13, scale = 2, nullable = false)
-    BigDecimal price;
+    Double price;
 
     @Builder.Default
     @Enumerated(EnumType.ORDINAL)
